Making it where contextMenuOn can take multiple event handlers
In my particular usecase I needed to be able to right or left click.

My solution is to create the openMenuEvent variable as an array and then
if attrs.contextMenuOn exists (and is a string) to split it by comma.
Then, taking the openMenuEvent array, loop through the element.on
handler to create mutlipe event handlers.

In the future, it might be worth making it where an array can actually
be sent in the javascript. However, that might be overkill since, if
they have an array...it won't take much to make it into a comma
delimited string.