Small error handing to make sure that the event string isn't empty
It occured to me that if you sent something like: "click," that it would
try to create an event of "" which would be invalid.

Added an if statement around the event creation (within the loop) to
make sure that the array position openMenuEvent[i] exists before trying
to create the event.